==> 5_MRPC/2_MPRC.R <==
library(GeneNet)
library(pcalg) 
library(MRPC)
library(stringr)
library(igraph)
set.seed(120)
library(ppcor)
library(Rgraphviz)
library(psych)

#import the new functions I have created
source("/mnt/iusers01/bk01/m20349jk/causal_hypertension/eQTL/pval_01/MRPC/MRdualPC.R")
#get the list of modules I have created using clustering
setwd("/mnt/iusers01/bk01/m20349jk/causal_hypertension/eQTL/pval_01/MRPC/nobatch")
col <- list.dirs(recursive = F)

ptm <- proc.time()
for(i in col){
	setwd("/mnt/iusers01/bk01/m20349jk/causal_hypertension/eQTL/pval_01/MRPC/nobatch")
	setwd(i)
	print(paste0("Start ", i))
	load("MRPC.Rdata")
	p <- length(labels)
	print(paste0("There are ", p, " nodes"))
	GV = sum(str_count(colnames(MRPC_data), pattern = "_b37"))

	#skel_pcalg <- dual_pc(cor_mat = suffStat_C1$C, N = suffStat_C1$n, dat = MRPC_data, alpha = 0.05, ord_ind = TRUE, skeleton = TRUE, pattern_graph = FALSE, max_ord = NULL)
	start_time <- Sys.time()
	filtered <- find_filtered(MRPC_data, 0.05, cor_mat = suffStat_C1$C)
	end_time <- Sys.time()
	print("Initial filtering : ")
	print(end_time - start_time)
	start_time <- Sys.time()
	skel_pcalg <- ModiSkeleton1(data = MRPC_data, filtered = filtered, suffStat = suffStat_C1, FDR= 0.05, alpha= 0.05, indepTest = "gaussCItest", FDRcontrol = "none", labels = labels, p = p, method = "stable")
	save(skel_pcalg, file = "skel_pcalg.Rdata")
	end_time <- Sys.time()
	print("Building skeleton : ")
	print(end_time - start_time)
	rm(MRPC_data)
	start_time <- Sys.time()
	MRPC.fit <- EdgeOrientation(skel_pcalg, GV, suffStat_C1, FDR = 0.05, alpha = 0.05, indepTest = 'gaussCItest', tau = 0.5, lambda = 0.25, FDRcontrol = "none", verbose = FALSE)
	save(MRPC.fit, file = "MRPC.fit.Rdata")
	end_time <- Sys.time()
	print(paste0("Edge Orientation : ", end_time - start_time))
	graph <- MRPC.fit@graph
	start_time <- Sys.time()
	adjacency <- as(graph, "matrix")
	write.csv(adjacency, file = "MRPC_adjacency.csv")
	defAttrs <- getDefaultAttrs()
	folder <- sub('./', '', i)
	pdf(paste0("MRPC_", folder,".pdf", sep =""), height = 30, width = 30) 
	plot(graph, attrs=list(node=list(fontsize=40)))
	dev.off() 

	remove <- nodes(graph)[-c(GV+1:length(nodes(graph)))]
	temp <- removeNode(remove, graph)
	defAttrs <- getDefaultAttrs()
	pdf(paste0("MRPC_", folder,"_noSNP.pdf"), height = 30, width = 30) 
	plot(temp, attrs=list(node=list(fontsize=40)))
	dev.off()

	igraph <- graph_from_graphnel(temp, name = TRUE, weight = TRUE, unlist.attrs = TRUE)
	d_search = bfs(igraph,"HYPERTENSION",neimode="in", unreachable=FALSE, order=TRUE, dist=TRUE)
	ancestors<-names(d_search$dist[!is.na(d_search$dist)])
	nodes <- nodes(MRPC.fit@graph)
	remove_all <- nodes[!(nodes %in% ancestors)]
	temp <- removeNode(remove_all, MRPC.fit@graph)
	defAttrs <- getDefaultAttrs()
	pdf(paste0("MRPC_", folder,"_ancestors.pdf"), height = 30, width = 30) 
	plot (temp, attrs=list(node=list(fontsize=40)))
	dev.off() 
	parent_adjacency <- adjacency[ancestors,ancestors]
	write.csv(parent_adjacency, quote= F , file = paste0("MRPC_ancestor_adjacency_", folder, ".csv"))
	end_time <- Sys.time()
	print(paste0("Saving and plotting : ", end_time - start_time))
	rm(parent_adjacency)
	rm(adjacency)
	rm(suffStat_C1)
	rm(labels)
	print(paste0("completed ", i))
}

proc.time() - ptm
print("done")
