==> 5_MRPC/5_speed_test.R <==
library(GeneNet)
library(pcalg) 
library(MRPC)
library(stringr)
library(igraph)
set.seed(120)
library(ppcor)
library(Rgraphviz)
library(psych)

#import the new functions I have created
source("/mnt/iusers01/bk01/m20349jk/causal_hypertension/eQTL/pval_01/MRPC/MRdualPC.R")
setwd("/mnt/iusers01/bk01/m20349jk/causal_hypertension/eQTL/pval_01/MRPC/nobatch")
col <- list.dirs(recursive = F)
modules <- vector()
number <- vector()
for(i in col){
	setwd("/mnt/iusers01/bk01/m20349jk/causal_hypertension/eQTL/pval_01/MRPC/nobatch")
	setwd(i)
	folder <- sub('./', '', i)
	print(paste0("Start ", folder))
	adj1 <- read.csv("MRPC_adjacency.csv")
	modules <- c(modules, i)
	number <- c(number, nrow(adj1))
}
all <- data.frame(modules, number)
all <- all[all$number < 1000,]


for(i in 1:nrow(all)){
	mod <- all$modules[i]
	setwd("/mnt/iusers01/bk01/m20349jk/causal_hypertension/eQTL/pval_01/MRPC/nobatch/")
	setwd(mod)
	folder <- sub('./', '', mod)
	print(paste0("Running MRPC on ", folder, " module"))
	load("MRPC.Rdata")
	p <- length(labels)
	print(paste0("There are ", p, " nodes"))
	GV = sum(str_count(colnames(MRPC_data), pattern = "_b37"))
	full_start_time <- Sys.time()
	start_time <- Sys.time()
	skel_pcalg <- ModiSkeleton(data = MRPC_data, suffStat = suffStat_C1, FDR= 0.05, alpha= 0.05, indepTest = "gaussCItest", FDRcontrol = "none", labels = labels, p = p, method = "stable")
	end_time <- Sys.time()
	print("Building skeleton : ")
	print(end_time - start_time)
	rm(MRPC_data)
	start_time <- Sys.time()
	MRPC.fit <- EdgeOrientation(skel_pcalg, GV, suffStat_C1, FDR = 0.05, alpha = 0.05, indepTest = 'gaussCItest', tau = 0.5, lambda = 0.25, FDRcontrol = "none", verbose = FALSE)
	end_time <- Sys.time()
	print("Edge Orientation : ")
	print(end_time - start_time)
	full_end_time <- Sys.time()
	print("MRPC module full time:")
	print(full_end_time - full_start_time)
	print(as.numeric(full_end_time - full_start_time,units="secs"))
	rm(MRPC.fit)
	rm(skel_pcalg)

	load("MRPC.Rdata")
	full_start_time <- Sys.time()
	start_time <- Sys.time()
	filtered <- find_filtered(MRPC_data, 0.05, cor_mat = suffStat_C1$C)
	end_time <- Sys.time()
	print("Initial filtering : ")
	print(end_time - start_time)
	start_time <- Sys.time()
	skel_pcalg <- ModiSkeleton1(data = MRPC_data, filtered = filtered, suffStat = suffStat_C1, FDR= 0.05, alpha= 0.05, indepTest = "gaussCItest", FDRcontrol = "none", labels = labels, p = p, method = "stable")
	end_time <- Sys.time()
	print("Building skeleton : ")
	print(end_time - start_time)
	rm(MRPC_data)
	start_time <- Sys.time()
	MRPC.fit <- EdgeOrientation(skel_pcalg, GV, suffStat_C1, FDR = 0.05, alpha = 0.05, indepTest = 'gaussCItest', tau = 0.5, lambda = 0.25, FDRcontrol = "none", verbose = FALSE)
	end_time <- Sys.time()
	print("Edge Orientation : ")
	print(end_time - start_time)
	full_end_time <- Sys.time()
	print("dualMRPC module full time:")
	print(full_end_time - full_start_time)
	print(as.numeric(full_end_time - full_start_time,units="secs"))
	rm(MRPC.fit)
	rm(skel_pcalg)
	rm(MRPC_data)
}
